requirements.txt
asgiref==3.4.0
Django==2.2
django-jsonfield==1.4.1
djangorestframework==3.12.4
jsonfield==3.1.0
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
